\name{load.homologeneData}
\alias{load.homologeneData}
\title{Load homologene data}
\description{Returns the homologene data for species with data included in this package}
\usage{
load.homologeneData();
}
\value{Returns a list containing the homologene data, taxonomy information and homologene build number}
\author{Katie Houlahan}
\examples{
homologeneData <- load.homologeneData();
}
\keyword{IO}

